+++ SDL_demo/QuaternionScene3D.h
@@ -17,7 +17,7 @@ public:
 	virtual void rotate_x(float d_angle);
 	virtual void rotate_y(float d_angle);
 	virtual void rotate_z(float d_angle);
-	virtual void rotate_vector(Vector vector, double d_angle);
+	virtual void rotate_vector(Vector vector, float d_angle);
 	virtual void move(float dx, float dy, float dz);
 
 	QuaternionScene3D& operator=(const QuaternionScene3D& other);
